requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.7.1
pydantic==1.10.12
python-multipart==0.0.6
opencv-python==4.8.0.76
face_recognition==1.3.0
numpy==1.24.4
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
twilio==8.4.1    # optional - SMS/voice alerts
